Made new database class
