+++ lib/DatabaseServices/Database.dart
@@ -7,19 +7,7 @@ class DatabaseServices {
 
   final CollectionReference users =
       FirebaseFirestore.instance.collection('Users');
-  final CollectionReference live =
-      FirebaseFirestore.instance.collection('Live');
 
-  Future addRecordToLive(String name, String grade) {
-    return live
-        .add({
-          'Name': name,
-          'Class': grade,
-          'EntryTime': DateTime.now().toUtc().millisecondsSinceEpoch
-        })
-        .then((value) => print("User Added"))
-        .catchError((error) => print("Failed to add user: $error"));
-  }
 
   Future<String> returnDesignation() async {
     String returnValue;
